cloud/firestore_backup/requirements.txt
google-auth==2.27.0
requests==2.31.0
